Task 1. Search for error messages in a log file
In this task, you must navigate to the /home/analyst/logs directory and report on the error messages in the server_logs.txt file. You’ll do this by using grep to search the file and output only the entries that are for errors.

Navigate to the /home/analyst/logs directory.
Use grep to filter the server_logs.txt file, and return all lines containing the text string error.